Add support for new AsyncCache from caffeine 2.7.0
